# Buffer results in section.all() by page size