@glensc 2021-08-11
While timing is not important in this PR, I was curious any way so I measured

- before: 7.1s, 6.7s, 6.8s, 6.8s, 6.6s
- after: 6.9s, 6.3s, 6.6s, 6.2s, 6.5s